requirements.txt
colorama==0.4.4
Django==3.2.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
